Fix parsing of opening hours
We originally used the Places API, therefore the way we store opening
hours in the local db matches that. Now that we've switched to OSM,
their format for opening hours is different. It is cumbersome to say the
least so we won't change the format of our local db. It means we have to
parse data coming from OSM as well as data coming from our local db.